% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEntropy}
\alias{getEntropy}
\title{getEntropy.}
\usage{
getEntropy(
  seu_obj = NULL,
  features = NULL,
  expMat = "path_to_expMat",
  platform = NULL
)
}
\arguments{
\item{seu_obj}{A Seurat object containing RNA count data.}
}
\value{
A data frame with the \code{sample_id}, \code{platform}, and the entropy value of the RNA count matrix, rounded to three decimal places.
}
\description{
Computes the entropy of the RNA count matrix within a Seurat object.
}
\details{
Entropy is used to quantify the diversity or uniformity of gene expression across the dataset. This function calculates the entropy of the RNA count matrix, reflecting the distribution of gene expression levels.
A higher entropy value suggests a more uniform distribution across genes, while a lower value indicates concentration in a smaller number of genes.
This measure can provide insights into the complexity of the cellular composition and the heterogeneity within the sample.
}
