% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGlobalFDR}
\alias{getGlobalFDR}
\title{getGlobalFDR}
\usage{
getGlobalFDR(seu_obj, features = NULL)
}
\arguments{
\item{seu_obj}{A Seurat object containing the spatial data, Must have
unique path and platform identifiers in its metadata.}

\item{features}{An optional vector of features (gene names) to include in the FDR calculation. If NULL, all features in the object are used.}
}
\value{
A data frame with columns for sample_id, platform, and the
calculated mean FDR across the specified features.
}
\description{
Calculate Specificity as Global False Discovery Rate (FDR)
}
\details{
This function calculates the global FDR based on the proportion of
negative control or 'blank' barcodes within the dataset. It's designed
to provide a measure of specificity across the entire panel of probes.
}
