% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMeanExpression}
\alias{getMeanExpression}
\title{getMeanExpression.}
\usage{
getMeanExpression(seu_obj, features = NULL)
}
\arguments{
\item{seu_obj}{seurat object.}

\item{features}{Optional; a vector of gene identifiers for which to perform the calculation. If NULL, all features are used.}
}
\value{
A data frame with probe type (Gene or Control), the mean expression values, and the associated sample and platform identifiers.
}
\description{
getMeanExpression.
}
