% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMeanSignalRatio}
\alias{getMeanSignalRatio}
\title{getMeanSignalRatio.}
\usage{
getMeanSignalRatio(seu_obj, features = NULL)
}
\arguments{
\item{seu_obj}{A Seurat object.}

\item{features}{Optional; specifies which genes to include in the calculation. If NULL, all genes in the RNA assay are considered.}
}
\value{
A data frame with sample_id, platform, and the calculated mean log-ratio.
}
\description{
It calculates log-ratio of mean gene counts to mean neg probe counts.
}
