% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMECR}
\alias{getMECR}
\title{getMECR.}
\usage{
getMECR(seu_obj)
}
\arguments{
\item{seu_obj}{A seurat object.}
}
\value{
A data frame containing the \code{sample_id}, \code{platform}, and the computed
MECR value. The MECR value is rounded to three decimal places and represents
the average mutually exclusive co-expression rate across the selected markers.
}
\description{
getMECR.
}
