% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetrics.R
\name{getMetrics}
\alias{getMetrics}
\title{getMetrics.}
\usage{
getMetrics(
  path = NULL,
  mecr = TRUE,
  autocorr = TRUE,
  clusterSilhouette = TRUE,
  sparsity = TRUE,
  entropy = TRUE,
  returnSeurat = FALSE,
  sample_meta = NULL
)
}
\arguments{
\item{path}{path to the directory containing sample folders. Each folder should be a separate sample.}

\item{mecr}{Logical; Default is TRUE, calculates the mean expression contrast ratio (MECR).}

\item{autocorr}{Logical; Default is TRUE, calculates Moran's I autocorrelation statistic.}

\item{clusterSilhouette}{Logical; Default is TRUE, calculates cluster silhouette widths.}

\item{sparsity}{Logical; Default is TRUE, calculates sparsity of the data.}

\item{entropy}{Logical; Default is TRUE, calculates Shannon entropy.}

\item{returnSeurat}{Logical; Default is FALSE, if TRUE, it returns a list with the original Seurat objects along with the calculated metrics.}

\item{sample_meta}{An optional data frame specifying sample metadata. If not provided, the function attempts to infer this information from the directory structure.}
}
\value{
Depending on the value of \code{returnSeurat}, this function returns either a list containing both the original Seurat objects and the metrics data frame, or just the metrics data frame.
}
\description{
This function will take a path of where each folder is a sample.
}
\details{
This function processes a collection of spatial assay samples located in a specified path, where each folder corresponds to a separate sample.
It computes a range of spatial metrics, including mean expression, transcripts per cell, transcripts per area, signal-to-noise ratio, and more.
The function is highly flexible, allowing the user to specify which metrics to calculate.
Additionally, it can return the processed data in a Seurat object format if needed.
}
