% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSilhouetteWidth}
\alias{getSilhouetteWidth}
\title{getSilhouetteWidth.}
\usage{
getSilhouetteWidth(seu_obj)
}
\arguments{
\item{seu_obj}{A Seurat object.}
}
\value{
A data frame.
}
\description{
It calculates silhouette width as cluster evaluation.
}
