% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSparsity}
\alias{getSparsity}
\title{getSparsity.}
\usage{
getSparsity(seu_obj)
}
\arguments{
\item{seu_obj}{A Seurat object.}
}
\value{
A data frame.
}
\description{
It shows the sparsity (as a count or proportion) of a matrix.
}
\details{
For example, .99 sparsity means 99\% of the values are zero. Similarly, a sparsity of 0 means the matrix is fully dense.
}
