% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTxPerArea}
\alias{getTxPerArea}
\title{getTxPerArea.}
\usage{
getTxPerArea(seu_obj, features = NULL)
}
\arguments{
\item{seu_obj}{A Seurat object.}

\item{features}{Optional; a vector of gene identifiers for which to perform the calculation. If NULL, all features are used.}
}
\value{
A data frame containing sample_id, platform, and the average number of transcripts per um2
}
\description{
It calculates transcripts per um2.
}
