% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTxPerCell}
\alias{getTxPerCell}
\title{getTxPerCell.}
\usage{
getTxPerCell(seu_obj, features = NULL)
}
\arguments{
\item{seu_obj}{A Seurat object.}

\item{features}{Optional; a vector of features to include in the calculation; If NULL, all features are used.}
}
\value{
A data frame containing sample_id, platform, and the average number of transcripts per cell.
}
\description{
Calculates the average number of transcripts per cell for the given features.
}
\details{
Calculate Transcripts per Cell

Calculates the average number of transcripts per cell for the given features.

This metric helps in understanding the transcriptomic richness of cells in spatial transcriptomics datasets.
}
