% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTxPerNuc}
\alias{getTxPerNuc}
\title{getTxPerNuc}
\usage{
getTxPerNuc(seu_obj, features = NULL)
}
\arguments{
\item{seu_obj}{A Seurat object.}

\item{features}{Optional; a vector of gene identifiers for which to perform the calculation. If NULL, all features are used.}
}
\description{
getTxPerNuc
}
