% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readSpatial}
\alias{readSpatial}
\alias{readTxMeta}
\title{readSpatial.}
\usage{
readSpatial(sample_id, path, platform)

readTxMeta(path, platform)
}
\arguments{
\item{sample_id}{A unique identifier for the sample being read.}

\item{path}{The file path to the directory containing the data files.}

\item{platform}{A character string indicating the platform of the data, Must be one of "Xenium", "CosMx", or "MERSCOPE".}
}
\value{
A Seurat object containing the spatial data, with additional metadata and embeddings as applicable per platform.
}
\description{
readSpatial.

readTxMeta.
}
\details{
Regardless of platform, data is stored in an assay named "RNA" for convenient.

For each platform, This table will be used by subsequent functions.
}
\section{Functions}{
\itemize{
\item \code{readSpatial()}: outputs a seurat object with some common metadata e for downstream comparison.

\item \code{readTxMeta()}: reads in the transcript localization/metadata table.

}}
