% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotFractionTxInCell}
\alias{plotFractionTxInCell}
\title{plotFractionTxInCell.}
\usage{
plotFractionTxInCell(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A ggplot object.
}
\description{
plotFractionTxInCell.
}
