% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotMeanExpression}
\alias{plotMeanExpression}
\title{plotMeanExpression.}
\usage{
plotMeanExpression(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A ggplot object.
}
\description{
plotMeanExpression.
}
