% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotNMI}
\alias{plotNMI}
\title{plotNMI}
\usage{
plotNMI(cluster_metrics)
}
\arguments{
\item{cluster_metrics}{Data frame with \code{sample_id}, \code{value}, and \code{metric} columns, filtered for "NMI".}
}
\value{
A ggplot object showing NMI values for each sample.
}
\description{
Creates a scatter plot to visualize NMI values across samples, indicating the quality of clustering.
}
