% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotPanelSize}
\alias{plotPanelSize}
\title{plotPanelSize}
\usage{
plotPanelSize(df)
}
\arguments{
\item{df}{A data frame expected to contain \code{sample_id} and \code{value}, where \code{value} represents the panel size.}
}
\value{
A ggplot object visualizing the panel sizes of samples.
}
\description{
plotPanelSize
}
