% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotSampleLabel}
\alias{plotSampleLabel}
\title{plotSampleLabel}
\usage{
plotSampleLabel(sample_meta)
}
\arguments{
\item{sample_meta}{A data frame containing at least two columns: \code{sample_id} and \code{platform}.}
}
\value{
A ggplot object representing the sample labels colored by platform.
}
\description{
plotSampleLabel
}
