% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMaxRatio}
\alias{getMaxRatio}
\title{getMaxRatio.}
\usage{
getMaxRatio(seu_obj, features = NULL)
}
\arguments{
\item{seu_obj}{seurat object.}

\item{features}{Optional; a vector of gene identifiers for which to perform the calculation. If NULL, all features are used.}
}
\value{
A data frame.
}
\description{
getMaxRatio.
}
