% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMECR}
\alias{getMECR}
\title{getMECR.}
\usage{
getMECR(seu_obj)
}
\arguments{
\item{seu_obj}{A seurat object.}
}
\value{
A data frame containing the \code{sample_id}, \code{platform}, and the computed
MECR value. The MECR value is rounded to three decimal places and represents
the average mutually exclusive co-expression rate across the selected markers.
}
\description{
It calculates  Mutually Exclusive Co-expression Rate (MECR) Implementation.
}
\details{
Based on a predefined set of markers and their associated cell types, this function computes the rate
at which pairs of markers are expressed in mutually exclusive patterns within cells. The calculation is limited
to markers found in the dataset and can be adjusted to focus on a subset by specifying it. The result is a
single MECR value that quantifies the overall mutually exclusive co-expression pattern, rounded to three decimal places.
}
