% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readTxMeta}
\alias{readTxMeta}
\title{readTxMeta.}
\usage{
readTxMeta(path, platform)
}
\arguments{
\item{path}{The file path to the directory containing the data files.}

\item{platform}{A character string indicating the platform of the data, Must be one of "Xenium", "CosMx", or "MERSCOPE".}
}
\description{
readTxMeta.
}
\details{
For each platform, This table will be used by subsequent functions.
}
\section{Functions}{
\itemize{
\item \code{readTxMeta()}: reads in the transcript localization/metadata table.

}}
